refactor: modify ci/cd file
add docker login and push
