+++ src/main/java/com/example/solutionchallenge/app/auth/controller/OauthController.java
@@ -2,7 +2,7 @@ package com.example.solutionchallenge.app.auth.controller;
 
 import com.example.solutionchallenge.app.auth.domain.oauth2.kakao.KakaoApiService;
 import com.example.solutionchallenge.app.common.exception.ApiException;
-import com.example.solutionchallenge.app.common.constant.ErrorCode;
+import com.example.solutionchallenge.app.auth.domain.constant.ErrorCode;
 import com.example.solutionchallenge.app.auth.dto.request.OauthRequestDto;
 import com.example.solutionchallenge.app.auth.dto.response.OauthResponseDto;
 import com.example.solutionchallenge.app.user.dto.RefreshTokenResponseDto;
@@ -11,13 +11,12 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
 
-import static com.example.solutionchallenge.app.common.constant.ErrorCode.UNAUTHORIZED;
+import static com.example.solutionchallenge.app.auth.domain.constant.ErrorCode.UNAUTHORIZED;
 
 @RestController
 @RequiredArgsConstructor
@@ -40,35 +39,36 @@ public class OauthController {
         }
         return oauthResponseDto;
     }
-    @RequestMapping(value="/")
+
+    @RequestMapping(value = "/")
     public String index() {
         return "index";
     }
 
-    @RequestMapping(value="/login")
+    @RequestMapping(value = "/login")
     public String login(@RequestParam("code") String code) {
         String access_Token = kakao.getAccessToken(code);
         System.out.println("controller access_token : " + access_Token);
         return "index";
     }
-}
 
     // 리프레시 토큰으로 액세스토큰 재발급 받는 로직
     @PostMapping("/token/refresh")
     public RefreshTokenResponseDto tokenRefresh(HttpServletRequest request) {
         RefreshTokenResponseDto refreshTokenResponseDto = new RefreshTokenResponseDto();
         Cookie[] list = request.getCookies();
-        if(list == null) {
+        if (list == null) {
             throw new ApiException(ErrorCode.INVALID_REFRESH_TOKEN);
         }
 
-        Cookie refreshTokenCookie = Arrays.stream(list).filter(cookie -> cookie.getName().equals("refresh_token")).collect(Collectors.toList()).get(0);
+        Cookie refreshTokenCookie = Arrays.stream(list).filter(cookie -> cookie.getName().equals("refresh_token"))
+                .collect(Collectors.toList()).get(0);
 
-        if(refreshTokenCookie == null) {
+        if (refreshTokenCookie == null) {
             throw new ApiException(ErrorCode.INVALID_REFRESH_TOKEN);
         }
         String accessToken = oauthService.refreshAccessToken(refreshTokenCookie.getValue());
         refreshTokenResponseDto.setAccessToken(accessToken);
         return refreshTokenResponseDto;
     }
-
+}
