+++ src/main/java/com/example/solutionchallenge/app/auth/domain/jwt/JwtFilter.java
@@ -1,7 +1,7 @@
 package com.example.solutionchallenge.app.auth.domain.jwt;
 
 import com.example.solutionchallenge.app.common.exception.ApiException;
-import com.example.solutionchallenge.app.common.constant.ErrorCode;
+import com.example.solutionchallenge.app.auth.domain.constant.ErrorCode;
 import com.example.solutionchallenge.app.user.dto.UserDto;
 import com.example.solutionchallenge.app.user.model.UserPrincipal;
 import com.example.solutionchallenge.app.user.service.UserService;
@@ -27,30 +27,30 @@ public class JwtFilter extends GenericFilterBean {
 
 
     //Filter 에서 액세스토큰이 유효한지 확인 후 SecurityContext에 계정정보 저장
-    @Override
-    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
-
-        HttpServletRequest httpServletRequest = (HttpServletRequest) servletRequest;
-        System.out.println(((HttpServletRequest) servletRequest).getRequestURL());
-        logger.info("[JwtFilter] : " + httpServletRequest.getRequestURL().toString());
-        String jwt = resolveToken(httpServletRequest);
-        System.out.println("jwt: " + jwt);
-
-       if (StringUtils.hasText(jwt) && jwtTokenService.validateToken(jwt)) {
-            Long userId = Long.valueOf(jwtTokenService.getPayload(jwt)); // 토큰에 있는 userId 가져오기
-            UserDto user = userService.findById(userId); // userId로
-            if(user == null) {
-                throw new ApiException(ErrorCode.NOT_EXIST_USER);
-            }
-            UserDetails userDetails = UserPrincipal.create(user);
-            UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-            SecurityContextHolder.getContext().setAuthentication(authentication);
-        } else {
-            throw new ApiException(ErrorCode.INVALID_ACCESS_TOKEN);
-        }
-
-        filterChain.doFilter(servletRequest, servletResponse);
-    }
+//    @Override
+//    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
+//
+//        HttpServletRequest httpServletRequest = (HttpServletRequest) servletRequest;
+//        System.out.println(((HttpServletRequest) servletRequest).getRequestURL());
+//        logger.info("[JwtFilter] : " + httpServletRequest.getRequestURL().toString());
+//        String jwt = resolveToken(httpServletRequest);
+//        System.out.println("jwt: " + jwt);
+//
+//       if (StringUtils.hasText(jwt) && jwtTokenService.validateToken(jwt)) {
+//            Long userId = Long.valueOf(jwtTokenService.getPayload(jwt)); // 토큰에 있는 userId 가져오기
+//            UserDto user = userService.findById(userId); // userId로
+//            if(user == null) {
+//                throw new ApiException(ErrorCode.NOT_EXIST_USER);
+//            }
+//            UserDetails userDetails = UserPrincipal.create(user);
+//            UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
+//            SecurityContextHolder.getContext().setAuthentication(authentication);
+//        } else {
+//            throw new ApiException(ErrorCode.INVALID_ACCESS_TOKEN);
+//        }
+//
+//        filterChain.doFilter(servletRequest, servletResponse);
+//    }
 
     public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
         HttpServletRequest httpServletRequest = (HttpServletRequest) servletRequest;
